Changed the main menu to horizontal view
The main menu has been changed to a horizontal view. I would suggest
looking through the main activity java file to see what I have commented
out. I have tested it and it works but the progress rate (i.e. 11.11% or
whatever) is not shown because it has no place to go at the current
moment. If it is necessary, I'll find a way to put it in.

Also, each picture calls its own function right now but that can change
to make it more efficient later.